文档(doc): doc
add platform judgement methods docs
