--- src/checkDevice/demo/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import common from '../index';
-
-export default () => {
-  console.log(common);
-  return <div>123</div>;
-};
