--- src/checkDevice/index.md
@@ -1,13 +1,0 @@
----
-title: checkDevice
-order: 1000
-nav:
-  title: tools
-  path: /tools
-group:
-  title: 判断设备环境
-  path: /checkDevice
-  # order: 2
----
-
-<code src="./demo/index.tsx" />
