--- src/common/demo/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import common from '../index';
-
-export default () => {
-  const test = common;
-  console.log(test);
-  return <div>123</div>;
-};
