--- src/common/index.md
@@ -1,13 +1,0 @@
----
-title: common
-order: 8
-nav:
-  title: tools
-  path: /tools
-group:
-  title: 通用
-  path: /common
-  order: 8
----
-
-<code src="./demo/index.tsx" />
