--- src/floatUtil/demo/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import float from '../index';
-
-export default () => {
-  console.log(float.divide(1, 3));
-  return <div>123</div>;
-};
