--- src/floatUtil/index.md
@@ -1,13 +1,0 @@
----
-title: floatUtil
-order: 1000
-nav:
-  title: tools
-  path: /tools
-group:
-  title: 浮点数计算
-  path: /floatUtil
-  order: 4
----
-
-<code src="./demo/index.tsx" />
