--- src/stringUtil/demo/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import stringUtil from '../index';
-
-export default () => {
-  console.log(stringUtil);
-  return <div>123</div>;
-};
