--- src/stringUtil/index.md
@@ -1,13 +1,0 @@
----
-title: stringUtil
-order: 1000
-nav:
-  title: tools
-  path: /tools
-group:
-  title: 字符串校验
-  path: /stringUtil
-  order: 6
----
-
-<code src="./demo/index.tsx" />
