--- src/validator/demo/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import Validator from '../index';
-
-export default () => {
-  const test = Validator('email');
-  console.log(test);
-  return <div>123</div>;
-};
