--- src/validator/index.md
@@ -1,12 +1,0 @@
----
-title: validator
-nav:
-  title: tools
-  path: /tools
-group:
-  title: 查看设备
-  path: /validator
-  order: 3
----
-
-<code src="./demo/index.tsx" />
